requirements.txt
websockets==12.0
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.3
jinja2==3.1.3
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9
